Remove duplicated class introduced during rewrite
